Add content to rightside of feed container
Fill in the content in the rightside of the feed container with
pictures and text. Add  a picture for each pretend user.
Add filler text for each article in the announcements container
Add icons to each project section.
Change font size for most text to vw or vh.
Align sidebar more to the middle.